requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==2.2.14
django-redis==4.12.1
djangorestframework==3.11.0
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
